Upgraded the plugin to replace the engine according to the EngineReplacementDecision strategy.
Added ability to pass a JSON Map of parameters to the gremlin script which will be added to the bindings. Those values can then referred to as variables inside the script. Will reduce the number of classes generated and held in PermGen as well as speedup the execution.

Massive Refactoring & cleanup.